requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
django-filter==24.2
sqlparse==0.5.0
gunicorn==20.1.0
typing_extensions==4.11.0
mysqlclient==2.1.0